refactor: remove css theming and jwpltx from common
