+++ packages/ui-react/src/components/ShareButton/ShareButton.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { copyToClipboard } from '@jwp/ott-common/src/utils/dom';
 import useBreakpoint, { Breakpoint } from '@jwp/ott-hooks-react/src/useBreakpoint';
 
+import { copyToClipboard } from '../../utils/clipboard';
 import Check from '../../icons/Check';
 import Share from '../../icons/Share';
 import Button from '../Button/Button';
